pyproject.toml
[project]
name = "chatlas-telegram"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "chatlas>=0.3.0",
    "google-generativeai>=0.8.3",
    "python-dotenv>=1.0.1",
    "python-telegram-bot[job-queue]>=21.10",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]
